+++ src/app/src/components/DrawMap/EditPolygonModal.js
@@ -13,13 +13,11 @@ import {
     ModalBody,
 } from '@chakra-ui/react';
 import { CloudUploadIcon } from '@heroicons/react/outline';
+import { useDispatch } from 'react-redux';
+import { renamePolygon } from '../../store/mapSlice';
 
-export default function EditPolygonModal({
-    isOpen,
-    defaultLabel,
-    onSubmit,
-    onClose,
-}) {
+export default function EditPolygonModal({ isOpen, defaultLabel, onClose }) {
+    const dispatch = useDispatch();
     const [label, setLabel] = useState(defaultLabel);
 
     useEffect(() => setLabel(defaultLabel), [defaultLabel]);
@@ -51,7 +49,7 @@ export default function EditPolygonModal({
                             <Button
                                 variant='cta'
                                 onClick={() => {
-                                    onSubmit(label);
+                                    dispatch(renamePolygon(label));
                                     onClose();
                                 }}
                             >