+++ src/app/src/store/store.js
@@ -1,0 +1,9 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+import mapReducer from './mapSlice';
+
+export default configureStore({
+    reducer: {
+        map: mapReducer,
+    },
+});